# Conflicts:
#	contracts/interfaces/IUniswapV3Pair.sol
#	test/__snapshots__/UniswapV3Factory.spec.ts.snap
#	test/__snapshots__/UniswapV3Pair.spec.ts.snap

+++ contracts/test/TestUniswapV3Callee.sol
@@ -118,9 +118,7 @@ contract TestUniswapV3Callee is IUniswapV3MintCallback, IUniswapV3SwapCallback {
         address sender = abi.decode(data, (address));
 
         emit MintCallback(amount0Owed, amount1Owed);
-        if (amount0Owed > 0)
-            IERC20(IUniswapV3Pair(msg.sender).token0()).transferFrom(sender, msg.sender, uint256(amount0Owed));
-        if (amount1Owed > 0)
-            IERC20(IUniswapV3Pair(msg.sender).token1()).transferFrom(sender, msg.sender, uint256(amount1Owed));
+        if (amount0Owed > 0) IERC20(IUniswapV3Pair(msg.sender).token0()).transferFrom(sender, msg.sender, amount0Owed);
+        if (amount1Owed > 0) IERC20(IUniswapV3Pair(msg.sender).token1()).transferFrom(sender, msg.sender, amount1Owed);
     }
 }
